Add support for exporting Markdown articles as PDF. Refs #573
The method used in this commit might not be the best way to implement this
feature. Nevertheless, it actually builds upon a documented feature of rst2pdf
and simplifies the code, as only a single interface is used.

The main logic behind this commit is:

 1- Load the markdown source files and pandoc's markdown reader;
 2- Convert the markdown input into an HTML representation + metadata;
 3- Merge the metadata and text into an ReStructured Text string;
 4- Feed this new string into rst2pdf.

For this feature to work, the user must have xhtml2pdf installed. A bug in
rst2pdf (https://github.com/rst2pdf/rst2pdf/issues/521) prevents it from
finding the PyPDF version used by latest xhtml2pdf versions. Hence, we have to
monkeypatch xhtml2pdf to provide the class rst2pdf expects *prior* to importing
rst2pdf.